Conflicts:
	Makefile
	PgCommon.pm
	pg_wrapper
	t/150_tsearch_stemming.t

+++ t/150_tsearch_stemming.t
@@ -1,16 +1,20 @@
 # Check tsearch, and stemming with dynamic creation of .affix/.dict files
 
-use strict; 
+use strict;
 
 use lib 't';
 use TestLib;
 
 my $version = $MAJORS[-1];
 
-use Test::More tests => 33;
-
 use lib '/usr/share/postgresql-common';
 use PgCommon;
+
+use Test::More tests => $PgCommon::rpm ? 1 : 39;
+if ($PgCommon::rpm) {
+    pass 'tsearch dictionaries not handled by postgresql-common on RedHat';
+    exit;
+}
 
 # test pg_updatedicts
 unlink '/Library/Caches/postgresql/dicts/en_us.affix';
@@ -18,7 +22,7 @@ unlink '/Library/Caches/postgresql/dicts/en_us.dict';
 unlink "/usr/local/opt/postgresql-$version/share/tsearch_data/en_us.affix";
 unlink "/usr/local/opt/postgresql-$version/share/tsearch_data/en_us.dict";
 is ((exec_as 0, 'pg_updatedicts'), 0, 'pg_updatedicts succeeded');
-ok -f '/Library/Caches/postgresql/dicts/en_us.affix', 
+ok -f '/Library/Caches/postgresql/dicts/en_us.affix',
     'pg_updatedicts created en_us.affix';
 ok -f '/Library/Caches/postgresql/dicts/en_us.dict',
     'pg_updatedicts created en_us.dict';
@@ -28,7 +32,7 @@ ok -l "/usr/local/opt/postgresql-$version/share/tsearch_data/en_us.dict",
     "pg_updatedicts created $version en_us.dict symlink";
 
 # create cluster
-is ((system "pg_createcluster $version main --start >/dev/null"), 0, "pg_createcluster $_ main");
+is ((system "pg_createcluster $version main --start >/dev/null"), 0, "pg_createcluster $version main");
 
 # create DB with en_US text search configuration
 is_program_out 'postgres', 'createdb fts', 0, '';
@@ -56,36 +60,44 @@ is ((exec_as 'postgres', 'psql -qd fts -c "
   INSERT INTO stuff (text) VALUES (\'PostgreSQL rocks\');
   INSERT INTO stuff (text) VALUES (\'Linux rocks\');
   INSERT INTO stuff (text) VALUES (\'I am your father\'\'s nephew\'\'s former roommate\');
+  INSERT INTO stuff (text) VALUES (\'3 cafés\');
   "'), 0, 'creating data table and search index');
 
 # test stemming
-is_program_out 'postgres', 'psql -Atd fts -c "
-    SELECT dictionary, lexemes FROM ts_debug(\'public.sc_english\', \'friendliest\')
-    "', 0, "english_ispell|{friendly}\n", 'stem search of correct word';
-is_program_out 'postgres', 'psql -Atd fts -c "
-    SELECT dictionary, lexemes FROM ts_debug(\'public.sc_english\', \'father\'\'s\')
-    "', 0, "english_ispell|{father}\n|\nenglish_ispell|{}\n", 'stem search of correct word';
-is_program_out 'postgres', 'psql -Atd fts -c "
-    SELECT dictionary, lexemes FROM ts_debug(\'public.sc_english\', \'duffles\')
-    "', 0, "english_stem|{duffl}\n", 'stem search of unknown word';
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT dictionary, lexemes FROM ts_debug(\'public.sc_english\', \'friendliest\')"',
+    0, "english_ispell|{friendly}\n", 'stem search of correct word';
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT dictionary, lexemes FROM ts_debug(\'public.sc_english\', \'father\'\'s\')"',
+    0, "english_ispell|{father}\n|\nenglish_ispell|{}\n", 'stem search of correct word';
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT dictionary, lexemes FROM ts_debug(\'public.sc_english\', \'duffles\')"',
+    0, "english_stem|{duffl}\n", 'stem search of unknown word';
 
 # test searching
-is_program_out 'postgres', 'psql -Atd fts -c "SELECT text 
-    FROM stuff, to_tsquery(\'rocks\') query 
-    WHERE query @@ to_tsvector(text)"', 0, 
-    "PostgreSQL rocks\nLinux rocks\n", 
-    'full text search, exact word';
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT text FROM stuff, to_tsquery(\'rocks\') query WHERE query @@ to_tsvector(text)"',
+    0, "PostgreSQL rocks\nLinux rocks\n", 'full text search, exact word';
 
-is_program_out 'postgres', 'psql -Atd fts -c "SELECT text 
-    FROM stuff, to_tsquery(\'rock\') query 
-    WHERE query @@ to_tsvector(text)"', 0, 
-    "PostgreSQL rocks\nLinux rocks\n", 
-    'full text search for word stem';
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT text FROM stuff, to_tsquery(\'rock\') query WHERE query @@ to_tsvector(text)"',
+    0, "PostgreSQL rocks\nLinux rocks\n", 'full text search for word stem';
 
-is_program_out 'postgres', 'psql -Atd fts -c "SELECT text 
-    FROM stuff, to_tsquery(\'roc\') query 
-    WHERE query @@ to_tsvector(text)"', 0, '',
-    'full text search for word substring fails';
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT text FROM stuff, to_tsquery(\'roc\') query WHERE query @@ to_tsvector(text)"',
+    0, '', 'full text search for word substring fails';
+
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT text FROM stuff, to_tsquery(\'cafés\') query WHERE query @@ to_tsvector(text)"',
+    0, "3 cafés\n", 'full text search, exact unicode word';
+
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT text FROM stuff, to_tsquery(\'café\') query WHERE query @@ to_tsvector(text)"',
+    0, "3 cafés\n", 'full text search for unicode word stem';
+
+is_program_out 'postgres',
+    'psql -Atd fts -c "SELECT text FROM stuff, to_tsquery(\'afé\') query WHERE query @@ to_tsvector(text)"',
+    0, '', 'full text search for unicode word substring fails';
 
 # clean up
 is ((system "pg_dropcluster $version main --stop"), 0);
